refactor(test): reorganizar la estrutura del projecto y mover java files a sus packages
