+++ src/test/java/com/coworking/resources/controller/reservation/ReservationControllerTest.java
@@ -1,4 +1,4 @@
-package com.coworking.resources.controller;
+package com.coworking.resources.controller.reservation;
 
 
 import com.coworking.reservation.controller.ReservationController;
